refactor(rolls): extract stress-roll module
Move executeStressRoll, buildPenaltyNote, SKILL_NAMES, and stress
helpers out of roll-executor.ts into stress-roll.ts. Re-export from
roll-executor.ts for backward compat.

refs #559
